Allow grouping input widgets
Add inputWidgetGroups array to configuration file that allows grouping of input widgets. The order of the members in each group defines the order of the widgets.
The entire widget group can be specified by adding "_widgets1", "_widgets2" etc to the overwriteSheetOrder$input array. Note that the position of the widget group inside the inputWidgetGroups array should be identical to the suffix of the group id (e.g. first group in array should be "_widgets1" in the overwriteSheetOrder array)