Update AutoNumeric to 4.6.0
Update AutoNumeric to the latest version 4.6.0. This version works with ES6 classes instead of as a jquery plugin.
With this update, invalid values (outside the min/max range) are no longer blocked, but instead the border color of the input field changes to be orange to indicate that the value entered is invalid. This should cause less confusion.