Implement external datasource option: customFunction
With the new remote datasource option: customFunction it is possible to place custom R code inside the renderer_<modelName> directory to import/export data. When importing data a function with the signature importCustom(symName, ...) will be called and is expected to return a data frame that has the required number and types of columns for the symbol.
When exporting data using customFunction, a function with the signature exportCustom(symName, dataToExport, ...) will be called. This feature is currently hidden/not documented.