Integrate new xlsio class into MIRO
When importing an Excel spreadsheet, you now get the option to select the index range. An index with the worksheet name _index will be automatically detected. Some first parsing of the index sheet is done before the user is able to click "import". Warnings and error will be displayed in local import dialog right after uploading an Excel spreadsheet.