Custom renderer editor: Avoid crashing when error in observer
When an error occurs in an observer, Shiny automatically crashes the session (see also: https://github.com/rstudio/shiny/issues/2904). This causes MIRO to crash when a user edits a custom renderer that throws an error in an observer. To prevent this, we redefine the observer function in the environment the custom renderer code is evaluated in.