MIRO Server admin panel: Adjust to new database schema
When registering a new app via the MIRO Server admin panel, a new postgres user as well as a schema with the same name as the user (M_{appName.toUpper()}) is now created. This will ensure that the database user of an app does not have access to data of other apps.