Scalars are now stored in separate tables
Scalars are now stored in separate tables instead of a single table with "scalar", "description", "value" column. When launching an app for the first time, views and triggers are created to allow accessing all input/output scalars at once. Thus, reading/writing of all input/output scalars can be accomplished with a single request.
The database class transforms scalar tables to the old format (scalar,description,value) so that much of the code in MIRO remains untouched (we might want to change that at some point in the future).
Scalar descriptions are not stored in the database any longer.

This change requires a lot of logic in the database migration script. Most of the migration can be automated but since we didnt store scalar types (set/parameter) in the database before, this has to be done manually. To accomplish this, the migration assistant was extended to allow type casting of scalar tables (tables with single dimension).

User groups will now always be stored with leading slash _. Usernames and group names must not start with a slash, which is enforced already. This makes it possible to visualize users and groups differently when modifying access permissions.

Double sliders and date ranges are now stored as <name of DDPar>_lo/<name of DDPar>_up instead of <name of DDPar>$lo/<name of DDPar>$up. This avoids a lot of conversion logic and is more extendable in the future where we may want to allow assigning distinct names of double dash parameters or even scalars to the lower and upper bound of those widgets.