Don't try to remove job results of discarded job due to race condition
When trying to remove job results of a discarded job, it is likely that the job is still in the "cancelling" state and thus, we get a 403 error when trying to remove the job results. We need to implement a flag in Engine that can be sent along the interrupt request to remove job results as soon as the job is cancelled.