MIRO Pivot: map custom colors to specific headers
When assigning custom colors, we now assign them to a particular header. This means that if there are UELs in the data that have no color assigned, they will be rendered with the default color (currently, greyish) and the custom colors will still be assigned to the specific headers they were assigned to.