Add check if scenario has already been solved
When solving a scenario, MIRO now informs if the scenario has already been solved before.
This is done by computing a SHA-256 hash over the input data + attachments with execute permissions.
When saving a scenario, the hash is also saved. When saving a scenario under a new name, the hash is transferred, unless:
 - scenario attachments are discarded and there were attachments with execute permissions
 - scenario is marked dirty