Merge Hypercube and Base Mode
We can now submit and import Hypercube jobs from the Base Mode when there is an Engine backend.
When submitting a new HC job, the input data is stored in the database.
The inex file for HC jobs only includes output data.
When importing a HC job, only the output data is imported.
When loading HC job into sandbox and saving it, the input data of the HC job is duplicated.
The input data of a HC job is removed automatically if there are no more scenarios that rely on it.