[Main] Make sure session is cleaned up when crashed
The cleanup code to cleanly exit a running session was located the end of the server function.
If anything goes wrong before this code is reached
(e.g. crash in custom renderer) the code is never ran.
Instead, the cleanup code is now at the beginning of the server function
so that the session is cleaned up properly even
if there is a crash later on in the server function.