[Meta] Support signed MIRO apps
MIRO apps can now be signed with an RSA private key.
If an app is unsiged, a user now gets a warning when importing it.
If an app is signed, but the signature is unknown, the user is warned
as well but has the option to trust the developer.
This will import the public key to the user's list of trusted keys and
future apps from this developer will be imported without warning.