[MIRO Server] Improve concurrent admin panels
When there are multiple people using the MIRO Server admin panel at
the same time, they could overwrite each other's configuration.
Now, when updating the specs.yaml file, we take into consideration
changes that happened between opening the admin panel and the time
the changes are to be written to disk.
There are still race conditions, but they are less likely to occur
now.