[Server] Fix issue migrating multiple databases per session
Since we reused the same ID for the shiny module to migrate db, when
opening the dialog for the second time, button observers would fire.
This resulted in the user not being able to modify the migration
configuration as the "submit" button would fire immediatly.
We now use a new unique ID each time the migration module is initiated.