[Views] Support global views
Views can now be registered globally (for each scenario) for an app.
This is achieved by placing a file "views.json" inside the
conf_<modelName> directory. The structure of this file is the same as
for scenario-specific/local views.
The Views$getIds() and Views$get() method gain a new "type" argument
that can be one of "global", "local" or "all" where "all" is the
default.
When "all" is specified, local views shadow global views.