[MIRO Pivot] Fix edit view buttons not visible
When overwriting a global view with a new local view, the buttons to
edit/remove the new view were not visible. We now always re-render the
whole views list when adding a new view, which also fixes an issue
where new views were not sorted properly (they were always appended at
the end).