[Init] Fix issue with integers as column headers
When validating data contract, we use the V8 implementation of
JSON.parse to parse the data contract JSON file.
However, this implementation does not keep the order of the headers,
which are implemented as an object instead of an array (which
unfortunately, we can't change anymore due to backward compatibility).
Thus, we now use the jsonlite implementation (R) which seems to keep
the order of the object when parsing the JSON file (even though I
could not find this stated somewhere explicitly).
This has the disadvantage that we have to replace defaults manually
(because we are not using the object modified by AJV).
Fortunately, the data contract only has a single default value.