[Meta] Validate app ID
We now enforce the app ID to be no longer than 60 characters, to only
contain lowercase ASCII characters, digits, , `_` or `-` as well as to
start with a lowercase ASCII character or digit.
This check is performed both when deploying app as well as when
importing it (both MIRO Desktop as well as MIRO Server).