[Auth] Update dependencies
This includes updating pydantic to version 2.
In the process, I refactored the code to use modern Python features.
This means the minimum Python version for the auth container is now
3.10.
Also fixed some issues in auth container that caused ugly exceptions
when Engine was irresponsive due to bad argument names.