[Log] Fix empty log output with GAMSPy models
Due to an issue in the GAMS control API, line endings on Windows are represented as `\r\r\n` when redirecting output from GAMS runs via `sys.stdout`. This caused MIRO to display only empty lines. We now ignore carriage returns (do not delete current line) not only if the next character is a newline but also if its another carriage return.