[Server] Fix reconnecting to apps did not work
We can not use the hidden div with data attribute approach any longer since shinyproxy now uses JSON strings to communicate some data.
Thus, we have to fall back to inline javascript, which violates our content security policy.
Thus, we have to use a nonce for the inline js and allow it in the CSP header.