[MIRO Pivot] Fix renderer triggered too often
When using MIRO Pivot with filter dropdowns that contain more than 500 elements, MIRO automatically uses server-side selectize dropdowns. This causes the renderer to trigger twice since the dropdown is first initialized without options and options are later loaded from the server. Thus, Shiny first sends the empty string to MIRO and then (once the dropdown is initialized) the selected (default) filters.
To work around this issue, we initialize the dropdown with choices and selected values set to `selected`.