[UI] Remove pivot renderer
The long deprecated renderer `pivot` base on the `rPivotTable` package was removed since the `miropivot` renderer is much seperior and the old one is no longer maintained for several years. This is a breaking change but should not affect too many people as it was removed from the configuration mode many years ago and is not documented either.