[CI] Automatically fix length-1-string-arrays in schema
R is treating arrays of strings with length 1 as regular strings (i.e. there is no difference between `c("asd")` and `"asd"`. This has consequences when launching the configuration mode as any length-1-string-arrays are converted to regular strings with jsonlite's writeJSON function. This means we have to adjust the `config_schema.json` schema to always allow strings as well as length-1-string-arrays.