[Server] Clean up app files on registration error
When something goes wrong while uploading an app, app files were not cleaned up resulting in a bad state where the app existed on the file system but not in the configuration.
This lead to users not being able to add an app with this id but also left them no way to remove it from the interface.