[CustomDataIO] Validate scalar input table
Allow providing `_scalars` table without description column as well as with missing scalars (not recommended). Providing invalid scalars throws a warning in the log.
The scalar validation logic was moved to the LocalFileIO class as e.g. XlsIO also requires parts of it. As a consequence, the CustomDataIO class now inherits from LocalFileIO as well.