[LocalFileIO] Relax validation of scalars df
When `fixScalarDf` is called with `ignoreInvalidScalars` set to `TRUE`, we no longer require scalar and value columns to exist, e.g. NULL would be a valid scalars DF and the template returned.
This fixes an issue with custom importers where scalars dataframe was validated even when no scalars were to be imported.