[Server] Fix file paths in Kubernetes
Instead of having separate directory trees for data/model directories, we now use only one treee where we have a `data` directory in the mounted file system.
Inside this `data` directory, we have a directory for each app (named by the `appId`) and inside this app-specific directory we have `model`, `data` and `cache`.
We mount these seperately in the UI containers where only the `cache` directory is writeable.