[Server] Auto-generate role prefix
In MIRO Server One, we set a static prefix `M_` to every role generated by admin panel. As we want to use the same Postgres cluster for multiple MIRO Server instances, we have to prefix roles with a common prefix for the same MIRO Server instance as roles are shared within a cluster.
For this purpose we create a table `sys_config` in the `public` schema, which is a key-value table where `role_prefix` is the key for the role prefix that is generated the first time an admin panel container is launched.