[UI] Clean up some unescaped embedded HTML
The JS function `Miro.confirmModalShow` did not escape any of its arguments by default,
potentially resulting in unescaped HTML being executed. Since the arguments are typically
language elements which are under our control (or control of the app developer), this is
not a major issue. However, it is just cleaner to escape by default and allow opting out
(e.g. in about dialog) if rendering HTML is explicitly wanted.